set up userRepository
