+++ apps/admin/src/pages/signUp/SignUpForm.tsx
@@ -1,21 +1,13 @@
+import { SignUp } from "@models";
 import { Form } from "../../components/Form/Form";
 import { FormField } from "../../components/Form/FormField";
+import { postSignUp } from "../../data/userRepository";
 
 export function SignUpForm() {
-    // https://github.com/rose-charlotte/EZLoc/issues/57
-    async function postUser(userInfo: SignUp) {
-        const res = await fetch(`${import.meta.env.VITE_API_ROUTE}user`, {
-            method: "POST",
-            headers: { "content-Type": "application/json" },
-            body: JSON.stringify(userInfo),
-        });
-        const mss = await res.json();
-        console.log("la reponse du back", mss);
-    }
     const onSubmit = (signUp: SignUp) => {
         console.log("submit", signUp);
 
-        postUser(signUp);
+        postSignUp(signUp);
     };
     return (
         <div className="mt-10 sm:mx-auto sm:w-full sm:max-w-sm">
@@ -29,14 +21,4 @@ export function SignUpForm() {
             </Form>
         </div>
     );
-}
-
-//A déplacer dans les models quand la PR SignIn sera passée
-export interface SignUp {
-    lastName: string;
-    firstName: string;
-    adress: string;
-    phone: string;
-    password: string;
-    passwordConfirm: string;
 }
