+++ models/SignUp.ts
@@ -1,0 +1,8 @@
+export interface SignUp {
+    lastName: string;
+    firstName: string;
+    adress: string;
+    phone: string;
+    password: string;
+    passwordConfirm: string;
+}
