+++ models/index.ts
@@ -1,1 +1,2 @@
 export * from "./SignIn";
+export * from "./SignUp";
